clean up DTO

+++ auth-service/src/main/java/com/goldatech/authservice/domain/dto/TransactionSummaryDTO.java
@@ -1,20 +1,18 @@
 package com.goldatech.authservice.domain.dto;
 
-
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
+import java.math.BigDecimal; // Import this
 
 @Data
 @NoArgsConstructor
 @AllArgsConstructor
 public class TransactionSummaryDTO {
-    // Add this as the first field to match the SELECT order
     private String email;
-
     private Long totalTransactions;
-    private Double totalSuccessAmount;
+    private BigDecimal totalSuccessAmount;
     private Long totalSuccessCount;
     private Long totalFailedCount;
-    private Double totalFailedAmount;
+    private BigDecimal totalFailedAmount;
 }
